+++ src/r0/context/get_context.rs
@@ -1,8 +1,11 @@
 //! [GET /_matrix/client/r0/rooms/{roomId}/context/{eventId}](https://matrix.org/docs/spec/client_server/r0.4.0.html#get-matrix-client-r0-rooms-roomid-context-eventid)
 
+use js_int::UInt;
 use ruma_api::ruma_api;
 use ruma_events::{collections::only, EventResult};
 use ruma_identifiers::{EventId, RoomId};
+
+use crate::r0::filter::RoomEventFilter;
 
 ruma_api! {
     metadata {
@@ -22,10 +25,15 @@ ruma_api! {
         ///
         /// Defaults to 10 if not supplied.
         #[ruma_api(query)]
-        pub limit: u8,
+        #[serde(skip_serializing_if = "Option::is_none")]
+        pub limit: Option<UInt>,
         /// The room to get events from.
         #[ruma_api(path)]
         pub room_id: RoomId,
+        /// A RoomEventFilter to filter returned events with.
+        #[serde(skip_serializing_if = "Option::is_none")]
+        #[ruma_api(query)]
+        pub filter: Option<RoomEventFilter>,
     }
 
     response {